srcs/inventory-app/requirements.txt
flask
requests
pika
python-dotenv
psycopg2-binary
